Leftovers from the refactory on components.
Refs: #109
